+++ cli/src/commands/mod.rs
@@ -1,11 +1,13 @@
 mod app;
 mod check;
+mod env;
 mod preset;
 mod self_install;
 mod shell;
 
 pub(crate) use app::*;
 pub(crate) use check::*;
+pub(crate) use env::*;
 pub(crate) use preset::*;
 pub(crate) use self_install::*;
 pub(crate) use shell::*;